backend/requirements.txt
pandas>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
